+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### BREAKING CHANGES
 
 ## [1.2.0] - 2019-10-15
+### Added
+- Add `url` property to config options.
+
 ### Changed
 - Upgrade mercury version and define it as peer dependency
 - Upgrade devDependencies